refactor(agents): extract shared LLM plumbing into builder/agents/llm.py (#309)
Move the mode-neutral model-construction and usage-mining helpers out of the
agent-mode agent_loop.py into a new builder/agents/llm.py so both build modes
depend on a shared module, instead of the deterministic pipeline importing from
the ReAct loop.

Moved byte-for-byte (via AST) so behaviour is identical: _build_chat_model,
_detect_provider, _get_request_timeout (+_DEFAULT_REQUEST_TIMEOUT),
_is_openai_reasoning_model (+_OPENAI_REASONING_PREFIXES), _recursion_limit,
_extract_token_usage, _extract_model_name.

- leaves.py (pipeline) and eval/react_factory.py now import from
  builder.agents.llm, removing the wrong-direction dependency on the agent-mode
  module.
- agent_loop.py imports back only the 5 helpers its body still uses; drops the
  now-unused `import os` and the relocated names from __all__.
- New tests/agents/test_llm.py pins every shared helper to the new import path
  (TDD) and smoke-tests behaviour; existing direct-import tests repointed.

First step of the build-mode harmonization (#309): one shared toolbox + llm
layer under both --interactive and --legacy-react. No behaviour change.
